requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.104.1
h11==0.14.0
idna==3.4
jose==1.0.0
mysql-connector==2.2.9
mysql-connector-python==8.2.0
mysql-connector-python-rf==2.2.2
passlib==1.7.4
protobuf==4.21.12
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2